@@ -32,12 +32,16 @@ public class UnlinkedFilesCrawler extends BackgroundTask<FileNodeViewModel> {
 
     private final Path directory;
     private final Filter<Path> fileFilter;
+    private final DateRange dateFilter;
+    private final ExternalFileSorter sorter;
     private final BibDatabaseContext databaseContext;
     private final FilePreferences filePreferences;
 
-    public UnlinkedFilesCrawler(Path directory, Filter<Path> fileFilter, BibDatabaseContext databaseContext, FilePreferences filePreferences) {
+    public UnlinkedFilesCrawler(Path directory, Filter<Path> fileFilter, DateRange dateFilter, ExternalFileSorter sorter, BibDatabaseContext databaseContext, FilePreferences filePreferences) {
         this.directory = directory;
         this.fileFilter = fileFilter;
+        this.dateFilter = dateFilter;
+        this.sorter = sorter;
         this.databaseContext = databaseContext;
         this.filePreferences = filePreferences;
     }
@@ -61,6 +65,9 @@ public class UnlinkedFilesCrawler extends BackgroundTask<FileNodeViewModel> {
      * For ensuring the capability to cancel the work of this recursive method, the first position in the integer array
      * 'state' must be set to 1, to keep the recursion running. When the states value changes, the method will resolve
      * its recursion and return what it has saved so far.
+     * <br>
+     * The files are filtered according to the {@link DateRange} filter value 
+     * and then sorted according to the {@link ExternalFileSorter} value.
      *
      * @throws IOException if directory is not a directory or empty
      */
@@ -92,11 +99,19 @@ public class UnlinkedFilesCrawler extends BackgroundTask<FileNodeViewModel> {
                 parent.getChildren().add(subRoot);
             }
         }
-
-        parent.setFileCount(files.size() + fileCount);
-        parent.getChildren().addAll(files.stream()
-                                         .map(FileNodeViewModel::new)
-                                         .collect(Collectors.toList()));
+        // filter files according to last edited date.
+        List<Path> filteredFiles = new ArrayList<Path>();
+        for (Path path : files) {
+            if (FileFilterUtils.filterByDate(path, dateFilter)) {
+                filteredFiles.add(path);
+            }
+        }
+        // sort files according to last edited date.
+        filteredFiles = FileFilterUtils.sortByDate(filteredFiles, sorter);
+        parent.setFileCount(filteredFiles.size() + fileCount);
+        parent.getChildren().addAll(filteredFiles.stream()
+                .map(FileNodeViewModel::new)
+                .collect(Collectors.toList()));
         return parent;
     }
 }
