@@ -1,5 +1,6 @@
 package org.jabref.gui.importer.fetcher;
 
+import java.util.Map;
 import java.util.SortedSet;
 
 import javafx.beans.property.ListProperty;
@@ -12,6 +13,7 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 
 import org.jabref.gui.DialogService;
+import org.jabref.gui.Globals;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.importer.ImportEntriesDialog;
 import org.jabref.gui.util.BackgroundTask;
@@ -23,6 +25,16 @@ import org.jabref.model.strings.StringUtil;
 import org.jabref.preferences.PreferencesService;
 
 import com.tobiasdiez.easybind.EasyBind;
+import de.saxsys.mvvmfx.utils.validation.FunctionBasedValidator;
+import de.saxsys.mvvmfx.utils.validation.ValidationMessage;
+import de.saxsys.mvvmfx.utils.validation.ValidationStatus;
+import de.saxsys.mvvmfx.utils.validation.Validator;
+import org.apache.lucene.queryparser.flexible.core.QueryNodeParseException;
+import org.apache.lucene.queryparser.flexible.core.parser.SyntaxParser;
+import org.apache.lucene.queryparser.flexible.standard.parser.ParseException;
+import org.apache.lucene.queryparser.flexible.standard.parser.StandardSyntaxParser;
+
+import static org.jabref.logic.importer.fetcher.transformers.AbstractQueryTransformer.NO_EXPLICIT_FIELD;
 
 public class WebSearchPaneViewModel {
 
@@ -31,6 +43,9 @@ public class WebSearchPaneViewModel {
     private final StringProperty query = new SimpleStringProperty();
     private final DialogService dialogService;
     private final StateManager stateManager;
+
+    private final Validator searchQueryValidator;
+    private final SyntaxParser parser = new StandardSyntaxParser();
 
     public WebSearchPaneViewModel(PreferencesService preferencesService, DialogService dialogService, StateManager stateManager) {
         this.dialogService = dialogService;
@@ -50,6 +65,29 @@ public class WebSearchPaneViewModel {
             int newIndex = fetchers.indexOf(newFetcher);
             preferencesService.storeSidePanePreferences(preferencesService.getSidePanePreferences().withWebSearchFetcherSelected(newIndex));
         });
+
+        searchQueryValidator = new FunctionBasedValidator<>(
+                query,
+                queryText -> {
+                    if (StringUtil.isBlank(queryText)) {
+                        // in case user did not enter something, it is treated as valid (to avoid UI WTFs)
+                        return null;
+                    }
+                    try {
+                        parser.parse(queryText, NO_EXPLICIT_FIELD);
+                        return null;
+                    } catch (ParseException e) {
+                        String element = e.currentToken.image;
+                        int position = e.currentToken.beginColumn;
+                        if (element == null) {
+                            return ValidationMessage.error(Localization.lang("Invalid query. Check position %0.", position));
+                        } else {
+                            return ValidationMessage.error(Localization.lang("Invalid query element '%0' at position %1", element, position));
+                        }
+                    } catch (QueryNodeParseException e) {
+                        return ValidationMessage.error("");
+                    }
+                });
     }
 
     public ObservableList<SearchBasedFetcher> getFetchers() {
@@ -77,25 +115,30 @@ public class WebSearchPaneViewModel {
     }
 
     public void search() {
-        if (StringUtil.isBlank(getQuery())) {
+        String query = getQuery().trim();
+        if (StringUtil.isBlank(query)) {
             dialogService.notify(Localization.lang("Please enter a search string"));
             return;
         }
-
         if (stateManager.getActiveDatabase().isEmpty()) {
             dialogService.notify(Localization.lang("Please open or start a new library before searching"));
             return;
         }
 
         SearchBasedFetcher activeFetcher = getSelectedFetcher();
+        Globals.getTelemetryClient().ifPresent(client -> client.trackEvent("search", Map.of("fetcher", activeFetcher.getName()), Map.of()));
 
         BackgroundTask<ParserResult> task;
-        task = BackgroundTask.wrap(() -> new ParserResult(activeFetcher.performSearch(getQuery().trim())))
-                             .withInitialMessage(Localization.lang("Processing %0", getQuery().trim()));
+        task = BackgroundTask.wrap(() -> new ParserResult(activeFetcher.performSearch(query)))
+                             .withInitialMessage(Localization.lang("Processing %0", query));
         task.onFailure(dialogService::showErrorDialogAndWait);
 
         ImportEntriesDialog dialog = new ImportEntriesDialog(stateManager.getActiveDatabase().get(), task);
         dialog.setTitle(activeFetcher.getName());
         dialogService.showCustomDialogAndWait(dialog);
+    }
+
+    public ValidationStatus queryValidationStatus() {
+        return searchQueryValidator.getValidationStatus();
     }
 }
