@@ -3,22 +3,28 @@ package org.jabref.logic.crawler;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Writer;
+import java.nio.charset.UnsupportedCharsetException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import org.jabref.logic.citationkeypattern.CitationKeyGenerator;
-import org.jabref.logic.crawler.git.GitHandler;
 import org.jabref.logic.database.DatabaseMerger;
+import org.jabref.logic.exporter.AtomicFileWriter;
 import org.jabref.logic.exporter.BibtexDatabaseWriter;
+import org.jabref.logic.exporter.SaveException;
 import org.jabref.logic.exporter.SavePreferences;
+import org.jabref.logic.git.SlrGitHandler;
 import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.OpenDatabase;
 import org.jabref.logic.importer.ParseException;
 import org.jabref.logic.importer.SearchBasedFetcher;
+import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.preferences.TimestampPreferences;
 import org.jabref.model.database.BibDatabase;
 import org.jabref.model.database.BibDatabaseContext;
@@ -42,15 +48,19 @@ import org.slf4j.LoggerFactory;
  * as well as the sharing, and versioning of results using git.
  */
 class StudyRepository {
-    // Tests work with study.bib
+    // Tests work with study.yml
     private static final String STUDY_DEFINITION_FILE_NAME = "study.yml";
     private static final Logger LOGGER = LoggerFactory.getLogger(StudyRepository.class);
     private static final Pattern MATCHCOLON = Pattern.compile(":");
     private static final Pattern MATCHILLEGALCHARACTERS = Pattern.compile("[^A-Za-z0-9_.\\s=-]");
+    // Currently we make assumptions about the configuration: the remotes, work and search branch names
+    private static final String REMOTE = "origin";
+    private static final String WORK_BRANCH = "work";
+    private static final String SEARCH_BRANCH = "search";
 
     private final Path repositoryPath;
     private final Path studyDefinitionFile;
-    private final GitHandler gitHandler;
+    private final SlrGitHandler gitHandler;
     private final Study study;
     private final ImportFormatPreferences importFormatPreferences;
     private final FileUpdateMonitor fileUpdateMonitor;
@@ -70,7 +80,7 @@ class StudyRepository {
      * @throws ParseException           Problem parsing the study definition file.
      */
     public StudyRepository(Path pathToRepository,
-                           GitHandler gitHandler,
+                           SlrGitHandler gitHandler,
                            ImportFormatPreferences importFormatPreferences,
                            FileUpdateMonitor fileUpdateMonitor,
                            SavePreferences savePreferences,
@@ -78,11 +88,6 @@ class StudyRepository {
                            BibEntryTypesManager bibEntryTypesManager) throws IOException, ParseException {
         this.repositoryPath = pathToRepository;
         this.gitHandler = gitHandler;
-        try {
-            gitHandler.updateLocalRepository();
-        } catch (GitAPIException e) {
-            LOGGER.error("Updating repository from remote failed");
-        }
         this.importFormatPreferences = importFormatPreferences;
         this.fileUpdateMonitor = fileUpdateMonitor;
         this.studyDefinitionFile = Path.of(repositoryPath.toString(), STUDY_DEFINITION_FILE_NAME);
@@ -92,40 +97,75 @@ class StudyRepository {
 
         if (Files.notExists(repositoryPath)) {
             throw new IOException("The given repository does not exists.");
-        } else if (Files.notExists(studyDefinitionFile)) {
+        }
+        try {
+            gitHandler.createCommitOnCurrentBranch("Save changes before searching.", false);
+            gitHandler.checkoutBranch(WORK_BRANCH);
+            updateWorkAndSearchBranch();
+        } catch (GitAPIException e) {
+            LOGGER.error("Could not checkout work branch");
+        }
+        if (Files.notExists(studyDefinitionFile)) {
             throw new IOException("The study definition file does not exist in the given repository.");
         }
         study = parseStudyFile();
-        this.setUpRepositoryStructure();
+        try {
+            // Update repository structure on work branch in case of changes
+            setUpRepositoryStructure();
+            gitHandler.createCommitOnCurrentBranch("Setup/Update Repository Structure", false);
+            gitHandler.checkoutBranch(SEARCH_BRANCH);
+            // If study definition does not exist on this branch or was changed on work branch, copy it from work
+            boolean studyDefinitionDoesNotExistOrChanged = !(Files.exists(studyDefinitionFile) && new StudyYamlParser().parseStudyYamlFile(studyDefinitionFile).equalsBesideLastSearchDate(study));
+            if (studyDefinitionDoesNotExistOrChanged) {
+                new StudyYamlParser().writeStudyYamlFile(study, studyDefinitionFile);
+            }
+            this.setUpRepositoryStructure();
+            gitHandler.createCommitOnCurrentBranch("Setup/Update Repository Structure", false);
+        } catch (GitAPIException e) {
+            LOGGER.error("Could not checkout search branch.");
+        }
+        try {
+            gitHandler.checkoutBranch(WORK_BRANCH);
+        } catch (GitAPIException e) {
+            LOGGER.error("Could not checkout work branch");
+        }
     }
 
     /**
      * Returns entries stored in the repository for a certain query and fetcher
      */
     public BibDatabaseContext getFetcherResultEntries(String query, String fetcherName) throws IOException {
-        return OpenDatabase.loadDatabase(getPathToFetcherResultFile(query, fetcherName), importFormatPreferences, timestampPreferences, fileUpdateMonitor).getDatabaseContext();
+        if (Files.exists(getPathToFetcherResultFile(query, fetcherName))) {
+            return OpenDatabase.loadDatabase(getPathToFetcherResultFile(query, fetcherName), importFormatPreferences, timestampPreferences, fileUpdateMonitor).getDatabaseContext();
+        }
+        return new BibDatabaseContext();
     }
 
     /**
      * Returns the merged entries stored in the repository for a certain query
      */
     public BibDatabaseContext getQueryResultEntries(String query) throws IOException {
-        return OpenDatabase.loadDatabase(getPathToQueryResultFile(query), importFormatPreferences, timestampPreferences, fileUpdateMonitor).getDatabaseContext();
+        if (Files.exists(getPathToQueryResultFile(query))) {
+            return OpenDatabase.loadDatabase(getPathToQueryResultFile(query), importFormatPreferences, timestampPreferences, fileUpdateMonitor).getDatabaseContext();
+        }
+        return new BibDatabaseContext();
     }
 
     /**
      * Returns the merged entries stored in the repository for all queries
      */
     public BibDatabaseContext getStudyResultEntries() throws IOException {
-        return OpenDatabase.loadDatabase(getPathToStudyResultFile(), importFormatPreferences, timestampPreferences, fileUpdateMonitor).getDatabaseContext();
+        if (Files.exists(getPathToStudyResultFile())) {
+            return OpenDatabase.loadDatabase(getPathToStudyResultFile(), importFormatPreferences, timestampPreferences, fileUpdateMonitor).getDatabaseContext();
+        }
+        return new BibDatabaseContext();
     }
 
     /**
      * The study definition file contains all the definitions of a study. This method extracts this study from the yaml study definition file
      *
      * @return Returns the BibEntries parsed from the study definition file.
-     * @throws IOException    Problem opening the input stream.
-     * @throws ParseException Problem parsing the study definition file.
+     * @throws IOException Problem opening the input stream.
      */
     private Study parseStudyFile() throws IOException {
         return new StudyYamlParser().parseStudyYamlFile(studyDefinitionFile);
@@ -160,20 +200,68 @@ class StudyRepository {
         return study;
     }
 
-    public void persist(List<QueryResult> crawlResults) throws IOException, GitAPIException {
-        try {
-            gitHandler.updateLocalRepository();
-        } catch (GitAPIException e) {
-            LOGGER.error("Updating repository from remote failed");
-        }
+    /**
+     * Persists the result locally and remotely by following the steps:
+     * Precondition: Currently checking out work branch
+     * <ol>
+     *     <li>Update the work and search branch</li>
+     *     <li>Persist the results on the search branch</li>
+     *     <li>Manually patch the diff of the search branch onto the work branch (as the merging will not work in
+     *     certain cases without a conflict as it is context sensitive. But for this use case we do not need it to be
+     *     context sensitive. So we can just prepend the patch without checking the "context" lines.</li>
+     *     <li>Update the remote tracking branches of the work and search branch</li>
+     * </ol>
+     */
+    public void persist(List<QueryResult> crawlResults) throws IOException, GitAPIException, SaveException {
+        updateWorkAndSearchBranch();
+        study.setLastSearchDate(LocalDate.now());
+        persistStudy();
+        gitHandler.createCommitOnCurrentBranch("Update search date", true);
+        gitHandler.checkoutBranch(SEARCH_BRANCH);
         persistResults(crawlResults);
         study.setLastSearchDate(LocalDate.now());
         persistStudy();
         try {
-            gitHandler.updateRemoteRepository("Conducted search " + LocalDate.now());
+            // First commit changes to search branch branch and update remote
+            String commitMessage = "Conducted search: " + LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS);
+            boolean newSearchResults = gitHandler.createCommitOnCurrentBranch(commitMessage, false);
+            gitHandler.checkoutBranch(WORK_BRANCH);
+            if (!newSearchResults) {
+                return;
+            }
+            // Patch new results into work branch
+            gitHandler.appendLatestSearchResultsOntoCurrentBranch(commitMessage + " - Patch", SEARCH_BRANCH);
+            // Update both remote tracked branches
+            updateRemoteSearchAndWorkBranch();
         } catch (GitAPIException e) {
-            LOGGER.error("Updating remote repository failed");
+            LOGGER.error("Updating remote repository failed", e);
         }
+    }
+
+    /**
+     * Update the remote tracking branches of the work and search branches
+     * The currently checked out branch is not changed if the method is executed successfully
+     */
+    private void updateRemoteSearchAndWorkBranch() throws IOException, GitAPIException {
+        String currentBranch = gitHandler.getCurrentlyCheckedOutBranch();
+        gitHandler.checkoutBranch(SEARCH_BRANCH);
+        gitHandler.pushCommitsToRemoteRepository();
+        gitHandler.checkoutBranch(WORK_BRANCH);
+        gitHandler.pushCommitsToRemoteRepository();
+        gitHandler.checkoutBranch(currentBranch);
+    }
+
+    /**
+     * Updates the local work and search branches with changes from their tracking remote branches
+     * The currently checked out branch is not changed if the method is executed successfully
+     */
+    private void updateWorkAndSearchBranch() throws IOException, GitAPIException {
+        String currentBranch = gitHandler.getCurrentlyCheckedOutBranch();
+        gitHandler.checkoutBranch(SEARCH_BRANCH);
+        gitHandler.pullOnCurrentBranch();
+        gitHandler.checkoutBranch(WORK_BRANCH);
+        gitHandler.pullOnCurrentBranch();
+        gitHandler.checkoutBranch(currentBranch);
     }
 
     private void persistStudy() throws IOException {
@@ -280,7 +368,7 @@ class StudyRepository {
      *
      * @param crawlResults The results that shall be persisted.
      */
-    private void persistResults(List<QueryResult> crawlResults) throws IOException {
+    private void persistResults(List<QueryResult> crawlResults) throws IOException, SaveException {
         DatabaseMerger merger = new DatabaseMerger(importFormatPreferences.getKeywordSeparator());
         BibDatabase newStudyResultEntries = new BibDatabase();
 
@@ -290,11 +378,12 @@ class StudyRepository {
                 BibDatabase fetcherEntries = fetcherResult.getFetchResult();
                 BibDatabaseContext existingFetcherResult = getFetcherResultEntries(result.getQuery(), fetcherResult.getFetcherName());
 
+                // Merge new entries into fetcher result file
+                merger.merge(existingFetcherResult.getDatabase(), fetcherEntries);
+
                 // Create citation keys for all entries that do not have one
                 generateCiteKeys(existingFetcherResult, fetcherEntries);
 
-                // Merge new entries into fetcher result file
-                merger.merge(existingFetcherResult.getDatabase(), fetcherEntries);
                 // Aggregate each fetcher result into the query result
                 merger.merge(queryResultEntries, fetcherEntries);
 
@@ -322,10 +411,21 @@ class StudyRepository {
         targetEntries.getEntries().stream().filter(bibEntry -> !bibEntry.hasCitationKey()).forEach(citationKeyGenerator::generateAndSetKey);
     }
 
-    private void writeResultToFile(Path pathToFile, BibDatabase entries) throws IOException {
+    private void writeResultToFile(Path pathToFile, BibDatabase entries) throws IOException, SaveException {
+        if (!Files.exists(pathToFile)) {
+            Files.createFile(pathToFile);
+        }
         try (Writer fileWriter = new FileWriter(pathToFile.toFile())) {
             BibtexDatabaseWriter databaseWriter = new BibtexDatabaseWriter(fileWriter, savePreferences, bibEntryTypesManager);
             databaseWriter.saveDatabase(new BibDatabaseContext(entries));
+        }
+        try (AtomicFileWriter fileWriter = new AtomicFileWriter(pathToFile, savePreferences.getEncoding(), savePreferences.shouldMakeBackup())) {
+            BibtexDatabaseWriter databaseWriter = new BibtexDatabaseWriter(fileWriter, savePreferences, bibEntryTypesManager);
+            databaseWriter.saveDatabase(new BibDatabaseContext(entries));
+        } catch (UnsupportedCharsetException ex) {
+            throw new SaveException(Localization.lang("Character encoding '%0' is not supported.", savePreferences.getEncoding().displayName()), ex);
+        } catch (IOException ex) {
+            throw new SaveException("Problems saving: " + ex, ex);
         }
     }
 
