@@ -28,6 +28,8 @@ import org.slf4j.LoggerFactory;
 
 /**
  * Utility class for trying to resolve URLs to full-text PDF for articles.
+ *
+ * Combines multiple {@link FulltextFetcher}s together. Each fetcher is invoked, the "best" result (sorted by the fetcher trust level) is returned.
  */
 public class FulltextFetchers {
     private static final Logger LOGGER = LoggerFactory.getLogger(FulltextFetchers.class);
@@ -59,7 +61,7 @@ public class FulltextFetchers {
         BibEntry clonedEntry = (BibEntry) entry.clone();
         Optional<DOI> doi = clonedEntry.getField(StandardField.DOI).flatMap(DOI::parse);
 
-        if (!doi.isPresent()) {
+        if (doi.isEmpty()) {
             findDoiForEntry(clonedEntry);
         }
 
