@@ -10,6 +10,7 @@ import org.jabref.logic.importer.WebFetchers;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.identifier.DOI;
+import org.jabref.model.strings.StringUtil;
 
 public class TitleFetcher implements IdBasedFetcher {
 
@@ -31,15 +32,19 @@ public class TitleFetcher implements IdBasedFetcher {
 
     @Override
     public Optional<BibEntry> performSearchById(String identifier) throws FetcherException {
+        if (StringUtil.isBlank(identifier)) {
+            return Optional.empty();
+        }
+
         BibEntry entry = new BibEntry();
         entry.setField(StandardField.TITLE, identifier);
+
         Optional<DOI> doi = WebFetchers.getIdFetcherForIdentifier(DOI.class).findIdentifier(entry);
-        if (!doi.isPresent()) {
+        if (doi.isEmpty()) {
             return Optional.empty();
         }
 
         DoiFetcher doiFetcher = new DoiFetcher(this.preferences);
-
         return doiFetcher.performSearchById(doi.get().getDOI());
     }
 }
