@@ -4,6 +4,8 @@ import java.util.Optional;
 import java.util.StringJoiner;
 import java.util.stream.Collectors;
 
+import org.jabref.model.strings.StringUtil;
+
 import org.apache.lucene.queryparser.flexible.core.nodes.BooleanQueryNode;
 import org.apache.lucene.queryparser.flexible.core.nodes.FieldQueryNode;
 import org.apache.lucene.queryparser.flexible.core.nodes.GroupQueryNode;
@@ -96,7 +98,7 @@ public abstract class AbstractQueryTransformer {
                 return s.isEmpty() ? Optional.empty() : Optional.of(s);
             }
             case NO_EXPLICIT_FIELD -> {
-                return Optional.of(handleUnFieldedTerm(term));
+                return handleUnFieldedTerm(term);
             }
             default -> {
                 // Just add unknown fields as default
@@ -184,21 +186,8 @@ public abstract class AbstractQueryTransformer {
      *
      * Default implementation: just return the term (in quotes if a space is contained)
      */
-    protected String handleUnFieldedTerm(String term) {
-        return quoteStringIfSpaceIsContained(term);
-    }
-
-    /**
-     * Encloses the given string with " if there is a space contained
-     *
-     * @return Returns a string
-     */
-    protected String quoteStringIfSpaceIsContained(String string) {
-        if (string.contains(" ")) {
-            return "\"" + string + "\"";
-        } else {
-            return string;
-        }
+    protected Optional<String> handleUnFieldedTerm(String term) {
+        return Optional.of(StringUtil.quoteStringIfSpaceIsContained(term));
     }
 
     protected String createKeyValuePair(String fieldAsString, String term) {
@@ -206,7 +195,7 @@ public abstract class AbstractQueryTransformer {
     }
 
     protected String createKeyValuePair(String fieldAsString, String term, String separator) {
-        return String.format("%s%s%s", fieldAsString, separator, quoteStringIfSpaceIsContained(term));
+        return String.format("%s%s%s", fieldAsString, separator, StringUtil.quoteStringIfSpaceIsContained(term));
     }
 
     /**
