@@ -1,5 +1,7 @@
 package org.jabref.logic.importer.fetcher.transformers;
 
+import org.jabref.model.strings.StringUtil;
+
 public class CollectionOfComputerScienceBibliographiesQueryTransformer extends AbstractQueryTransformer {
 
     @Override
@@ -29,7 +31,7 @@ public class CollectionOfComputerScienceBibliographiesQueryTransformer extends A
 
     @Override
     protected String handleJournal(String journalTitle) {
-        return quoteStringIfSpaceIsContained(journalTitle);
+        return StringUtil.quoteStringIfSpaceIsContained(journalTitle);
     }
 
     @Override
