@@ -1,5 +1,7 @@
 package org.jabref.logic.importer.fetcher.transformers;
 
+import org.jabref.model.strings.StringUtil;
+
 /**
  * Default query transformer without any boolean operators
  */
@@ -22,16 +24,16 @@ public class DefaultQueryTransformer extends YearAndYearRangeByFilteringQueryTra
 
     @Override
     protected String handleAuthor(String author) {
-        return quoteStringIfSpaceIsContained(author);
+        return StringUtil.quoteStringIfSpaceIsContained(author);
     }
 
     @Override
     protected String handleTitle(String title) {
-        return quoteStringIfSpaceIsContained(title);
+        return StringUtil.quoteStringIfSpaceIsContained(title);
     }
 
     @Override
     protected String handleJournal(String journalTitle) {
-        return quoteStringIfSpaceIsContained(journalTitle);
+        return StringUtil.quoteStringIfSpaceIsContained(journalTitle);
     }
 }
