@@ -24,6 +24,9 @@ class UpdateTimestampListenerTest {
     private PreferencesService preferencesMock;
     private TimestampPreferences timestampPreferencesMock;
 
+    private final String baseDate = "2000-1-1";
+    private final String newDate = "2000-1-2";
+
     @BeforeEach
     void setUp() {
         database = new BibDatabase();
@@ -39,9 +42,6 @@ class UpdateTimestampListenerTest {
 
     @Test
     void updateTimestampEnabled() {
-        final String baseDate = "2000-1-1";
-        final String newDate = "2000-1-2";
-
         final boolean includeTimestamp = true;
 
         when(timestampPreferencesMock.now()).thenReturn(newDate);
@@ -60,9 +60,6 @@ class UpdateTimestampListenerTest {
 
     @Test
     void updateTimestampDisabled() {
-        final String baseDate = "2000-1-1";
-        final String newDate = "2000-1-2";
-
         final boolean includeTimestamp = false;
 
         when(timestampPreferencesMock.now()).thenReturn(newDate);
