@@ -1,10 +1,14 @@
 package org.jabref.logic.bibtex;
 
 import java.nio.file.Path;
+import java.util.stream.Stream;
 
 import org.jabref.model.entry.LinkedFile;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -31,12 +35,19 @@ public class FileFieldWriterTest {
         assertNull(FileFieldWriter.quote(null));
     }
 
-    @Test
-    public void testEncodeStringArray() {
-        assertEquals("a:b;c:d", FileFieldWriter.encodeStringArray(new String[][] {{"a", "b"}, {"c", "d"}}));
-        assertEquals("a:;c:d", FileFieldWriter.encodeStringArray(new String[][] {{"a", ""}, {"c", "d"}}));
-        assertEquals("a:" + null + ";c:d", FileFieldWriter.encodeStringArray(new String[][] {{"a", null}, {"c", "d"}}));
-        assertEquals("a:\\:b;c\\;:d", FileFieldWriter.encodeStringArray(new String[][] {{"a", ":b"}, {"c;", "d"}}));
+    private static Stream<Arguments> getEncodingTestData() {
+        return Stream.of(
+                Arguments.of("a:b;c:d", new String[][] {{"a", "b"}, {"c", "d"}}),
+                Arguments.of("a:;c:d", new String[][] {{"a", ""}, {"c", "d"}}),
+                Arguments.of("a:" + null + ";c:d", new String[][] {{"a", null}, {"c", "d"}}),
+                Arguments.of("a:\\:b;c\\;:d", new String[][] {{"a", ":b"}, {"c;", "d"}})
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource("getEncodingTestData")
+    public void testEncodeStringArray(String expected, String[][] values) {
+        assertEquals(expected, FileFieldWriter.encodeStringArray(values));
     }
 
     @Test
