@@ -8,10 +8,9 @@ import java.util.Set;
 
 import org.jabref.logic.importer.fetcher.TrustLevel;
 import org.jabref.model.entry.BibEntry;
+import org.jabref.model.entry.field.StandardField;
 import org.jabref.testutils.category.FetcherTest;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -20,17 +19,8 @@ import static org.mockito.Mockito.when;
 
 @FetcherTest
 public class FulltextFetchersTest {
-    private BibEntry entry;
 
-    @BeforeEach
-    public void setUp() {
-        entry = new BibEntry();
-    }
-
-    @AfterEach
-    public void tearDown() {
-        entry = null;
-    }
+    private BibEntry entry = new BibEntry();
 
     @Test
     public void acceptPdfUrls() throws MalformedURLException {
@@ -60,17 +50,20 @@ public class FulltextFetchersTest {
 
     @Test
     public void higherTrustLevelWins() throws IOException, FetcherException {
-        final URL lowUrl = new URL("http://docs.oasis-open.org/opencsa/sca-bpel/sca-bpel-1.1-spec-cd-01.pdf");
-        final URL highUrl = new URL("http://docs.oasis-open.org/wsbpel/2.0/OS/wsbpel-v2.0-OS.pdf");
-
         FulltextFetcher finderHigh = mock(FulltextFetcher.class);
-        FulltextFetcher finderLow = mock(FulltextFetcher.class);
         when(finderHigh.getTrustLevel()).thenReturn(TrustLevel.SOURCE);
-        when(finderLow.getTrustLevel()).thenReturn(TrustLevel.UNKNOWN);
+        final URL highUrl = new URL("http://docs.oasis-open.org/wsbpel/2.0/OS/wsbpel-v2.0-OS.pdf");
         when(finderHigh.findFullText(entry)).thenReturn(Optional.of(highUrl));
+
+        FulltextFetcher finderLow = mock(FulltextFetcher.class);
+        when(finderLow.getTrustLevel()).thenReturn(TrustLevel.UNKNOWN);
+        final URL lowUrl = new URL("http://docs.oasis-open.org/opencsa/sca-bpel/sca-bpel-1.1-spec-cd-01.pdf");
         when(finderLow.findFullText(entry)).thenReturn(Optional.of(lowUrl));
 
         FulltextFetchers fetcher = new FulltextFetchers(Set.of(finderLow, finderHigh));
+
+        // set an (arbitrary) DOI to the test entry to skip side effects inside the "findFullTextPDF" method
+        entry.setField(StandardField.DOI, "10.5220/0007903201120130");
 
         assertEquals(Optional.of(highUrl), fetcher.findFullTextPDF(entry));
     }
