@@ -23,6 +23,23 @@ import static org.mockito.Mockito.when;
 @FetcherTest
 class IEEETest implements SearchBasedFetcherCapabilityTest, PagedSearchFetcherTest {
 
+    private final BibEntry IGOR_NEWCOMERS = new BibEntry(StandardEntryType.InProceedings)
+            .withField(StandardField.AUTHOR, "Igor Steinmacher and Tayana Uchoa Conte and Christoph Treude and Marco Aurélio Gerosa")
+            .withField(StandardField.DATE, "14-22 May 2016")
+            .withField(StandardField.YEAR, "2016")
+            .withField(StandardField.EVENTDATE, "14-22 May 2016")
+            .withField(StandardField.EVENTTITLEADDON, "Austin, TX, USA")
+            .withField(StandardField.LOCATION, "Austin, TX, USA")
+            .withField(StandardField.DOI, "10.1145/2884781.2884806")
+            .withField(StandardField.JOURNALTITLE, "2016 IEEE/ACM 38th International Conference on Software Engineering (ICSE)")
+            .withField(StandardField.PAGES, "273--284")
+            .withField(StandardField.ISBN, "978-1-5090-2071-3")
+            .withField(StandardField.ISSN, "1558-1225")
+            .withField(StandardField.PUBLISHER, "IEEE")
+            .withField(StandardField.KEYWORDS, "Portals, Documentation, Computer bugs, Joining processes, Industries, Open source software, Newcomers, Newbies, Novices, Beginners, Open Source Software, Barriers, Obstacles, Onboarding, Joining Process")
+            .withField(StandardField.TITLE, "Overcoming Open Source Project Entry Barriers with a Portal for Newcomers")
+            .withField(StandardField.FILE, ":https\\://ieeexplore.ieee.org/stamp/stamp.jsp?tp=&arnumber=7886910:PDF");
+
     private IEEE fetcher;
     private BibEntry entry;
 
@@ -105,26 +122,19 @@ class IEEETest implements SearchBasedFetcherCapabilityTest, PagedSearchFetcherTe
     }
 
     @Test
-    void searchByQueryFindsEntry() throws Exception {
-        BibEntry expected = new BibEntry(StandardEntryType.InProceedings)
-                .withField(StandardField.AUTHOR, "Igor Steinmacher and Tayana Uchoa Conte and Christoph Treude and Marco Aurélio Gerosa")
-                .withField(StandardField.DATE, "14-22 May 2016")
-                .withField(StandardField.YEAR, "2016")
-                .withField(StandardField.EVENTDATE, "14-22 May 2016")
-                .withField(StandardField.EVENTTITLEADDON, "Austin, TX")
-                .withField(StandardField.LOCATION, "Austin, TX")
-                .withField(StandardField.DOI, "10.1145/2884781.2884806")
-                .withField(StandardField.JOURNALTITLE, "2016 IEEE/ACM 38th International Conference on Software Engineering (ICSE)")
-                .withField(StandardField.PAGES, "273--284")
-                .withField(StandardField.ISBN, "978-1-5090-2071-3")
-                .withField(StandardField.ISSN, "1558-1225")
-                .withField(StandardField.PUBLISHER, "IEEE")
-                .withField(StandardField.KEYWORDS, "Portals, Documentation, Computer bugs, Joining processes, Industries, Open source software, Newcomers, Newbies, Novices, Beginners, Open Source Software, Barriers, Obstacles, Onboarding, Joining Process")
-                .withField(StandardField.TITLE, "Overcoming Open Source Project Entry Barriers with a Portal for Newcomers")
-                .withField(StandardField.FILE, ":https\\://ieeexplore.ieee.org/stamp/stamp.jsp?tp=&arnumber=7886910:PDF")
-                .withField(StandardField.ABSTRACT, "Community-based Open Source Software (OSS) projects are usually self-organized and dynamic, receiving contributions from distributed volunteers. Newcomer are important to the survival, long-term success, and continuity of these communities. However, newcomers face many barriers when making their first contribution to an OSS project, leading in many cases to dropouts. Therefore, a major challenge for OSS projects is to provide ways to support newcomers during their first contribution. In this paper, we propose and evaluate FLOSScoach, a portal created to support newcomers to OSS projects. FLOSScoach was designed based on a conceptual model of barriers created in our previous work. To evaluate the portal, we conducted a study with 65 students, relying on qualitative data from diaries, self-efficacy questionnaires, and the Technology Acceptance Model. The results indicate that FLOSScoach played an important role in guiding newcomers and in lowering barriers related to the orientation and contribution process, whereas it was not effective in lowering technical barriers. We also found that FLOSScoach is useful, easy to use, and increased newcomers' confidence to contribute. Our results can help project maintainers on deciding the points that need more attention in order to help OSS project newcomers overcome entry barriers.");
+    void searchByPlainQueryFindsEntry() throws Exception {
         List<BibEntry> fetchedEntries = fetcher.performSearch("Overcoming Open Source Project Entry Barriers with a Portal for Newcomers");
-        assertEquals(Collections.singletonList(expected), fetchedEntries);
+        // Abstract should not be included in JabRef tests
+        fetchedEntries.forEach(entry -> entry.clearField(StandardField.ABSTRACT));
+        assertEquals(Collections.singletonList(IGOR_NEWCOMERS), fetchedEntries);
+    }
+
+    @Test
+    void searchByQuotedQueryFindsEntry() throws Exception {
+        List<BibEntry> fetchedEntries = fetcher.performSearch("\"Overcoming Open Source Project Entry Barriers with a Portal for Newcomers\"");
+        // Abstract should not be included in JabRef tests
+        fetchedEntries.forEach(entry -> entry.clearField(StandardField.ABSTRACT));
+        assertEquals(Collections.singletonList(IGOR_NEWCOMERS), fetchedEntries);
     }
 
     @Override
