@@ -17,11 +17,10 @@ public class HTMLCharacterCheckerTest {
     private final BibEntry entry = new BibEntry();
 
     @Test
-    void fieldNullValueCheck() {
-        Exception exception = assertThrows(
+    void testSettingNullThrowsNPE() {
+        assertThrows(
                 NullPointerException.class,
-                () -> entry.setField(StandardField.AUTHOR, null),
-                "field value must not be null"
+                () -> entry.setField(StandardField.AUTHOR, null)
         );
     }
 
@@ -60,5 +59,4 @@ public class HTMLCharacterCheckerTest {
         entry.setField(StandardField.JOURNAL, "&Auml;rling Str&ouml;m for &#8211; &#x2031;");
         assertEquals(List.of(new IntegrityMessage("HTML encoded character found", entry, StandardField.JOURNAL)), checker.check(entry));
     }
-
 }
