@@ -14,16 +14,20 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class MergeReviewIntoCommentActionMigrationTest {
+
     private MergeReviewIntoCommentMigration action;
+    private BibEntry entry;
+    private BibEntry expectedEntry;
 
     @BeforeEach
     public void setUp() {
         action = new MergeReviewIntoCommentMigration();
+        entry = createMinimalBibEntry();
+        expectedEntry = createMinimalBibEntry();
     }
 
     @Test
     public void noFields() {
-        BibEntry entry = createMinimalBibEntry();
         ParserResult actualParserResult = new ParserResult(Collections.singletonList(entry));
 
         action.performMigration(actualParserResult);
@@ -33,11 +37,9 @@ class MergeReviewIntoCommentActionMigrationTest {
 
     @Test
     public void reviewField() {
-        BibEntry actualEntry = createMinimalBibEntry();
-        actualEntry.setField(StandardField.REVIEW, "My Review");
-        ParserResult actualParserResult = new ParserResult(Collections.singletonList(actualEntry));
+        entry.setField(StandardField.REVIEW, "My Review");
+        ParserResult actualParserResult = new ParserResult(Collections.singletonList(entry));
 
-        BibEntry expectedEntry = createMinimalBibEntry();
         expectedEntry.setField(StandardField.COMMENT, "My Review");
 
         action.performMigration(actualParserResult);
@@ -47,7 +49,6 @@ class MergeReviewIntoCommentActionMigrationTest {
 
     @Test
     public void commentField() {
-        BibEntry entry = createMinimalBibEntry();
         entry.setField(StandardField.COMMENT, "My Comment");
         ParserResult actualParserResult = new ParserResult(Collections.singletonList(entry));
 
@@ -60,11 +61,9 @@ class MergeReviewIntoCommentActionMigrationTest {
     public void multiLineReviewField() {
         String commentString = "My Review\n\nSecond Paragraph\n\nThird Paragraph";
 
-        BibEntry actualEntry = createMinimalBibEntry();
-        actualEntry.setField(StandardField.REVIEW, commentString);
-        ParserResult actualParserResult = new ParserResult(Collections.singletonList(actualEntry));
+        entry.setField(StandardField.REVIEW, commentString);
+        ParserResult actualParserResult = new ParserResult(Collections.singletonList(entry));
 
-        BibEntry expectedEntry = createMinimalBibEntry();
         expectedEntry.setField(StandardField.COMMENT, commentString);
 
         action.performMigration(actualParserResult);
@@ -75,13 +74,11 @@ class MergeReviewIntoCommentActionMigrationTest {
     @Test
     @Disabled("Re-enable if the MergeReviewIntoCommentMigration.mergeCommentFieldIfPresent() does not block and wait for user input.")
     public void reviewAndCommentField() {
-        BibEntry actualEntry = createMinimalBibEntry();
-        actualEntry.setField(StandardField.REVIEW, "My Review");
-        actualEntry.setField(StandardField.COMMENT, "My Comment");
+        entry.setField(StandardField.REVIEW, "My Review");
+        entry.setField(StandardField.COMMENT, "My Comment");
 
-        ParserResult actualParserResult = new ParserResult(Collections.singletonList(actualEntry));
+        ParserResult actualParserResult = new ParserResult(Collections.singletonList(entry));
 
-        BibEntry expectedEntry = createMinimalBibEntry();
         expectedEntry.setField(StandardField.COMMENT, "My Comment\n" + Localization.lang("Review") + ":\nMy Review");
 
         action.performMigration(actualParserResult);
