@@ -22,7 +22,6 @@ public class EntryLinkListTest {
     private ParsedEntryLink link;
     private BibEntry source;
     private BibEntry target;
-    private BibEntry entry;
 
     @BeforeEach
     public void before() {
@@ -34,8 +33,8 @@ public class EntryLinkListTest {
     }
 
     private BibEntry create(String citeKey) {
-        BibEntry entry = new BibEntry();
-        entry.setCitationKey(citeKey);
+        BibEntry entry = new BibEntry()
+                .withCitationKey(citeKey);
         database.insertEntry(entry);
         return entry;
     }
@@ -63,9 +62,8 @@ public class EntryLinkListTest {
 
     @Test
     public void givenBibEntryWhenParsingThenExpectLink() {
-        entry = create("entry");
-        ParsedEntryLink expected = new ParsedEntryLink(entry);
-        assertFalse(expected.getLinkedEntry().isEmpty());
+      ParsedEntryLink expected = new ParsedEntryLink(new BibEntry().withCitationKey("key"));
+      assertFalse(expected.getLinkedEntry().isEmpty());
     }
 
     @Test
