@@ -1,10 +1,12 @@
 package org.jabref.model.groups;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
+import org.jabref.model.FieldChange;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.search.matchers.AndMatcher;
@@ -109,7 +111,7 @@ public class GroupTreeNodeTest {
     }
 
     @BeforeEach
-    void setUp() throws Exception {
+    void setUp() {
         entries.clear();
         entry = new BibEntry();
         entries.add(entry);
@@ -160,12 +162,12 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void findMatchesReturnsEmptyForEmptyList() throws Exception {
+    void findMatchesReturnsEmptyForEmptyList() {
         assertEquals(Collections.emptyList(), getNodeInSimpleTree().findMatches(Collections.emptyList()));
     }
 
     @Test
-    void findMatchesOneEntry() throws Exception {
+    void findMatchesOneEntry() {
         GroupTreeNode parent = getNodeInSimpleTree();
         GroupTreeNode node = parent.addSubgroup(
                 new WordKeywordGroup("node", GroupHierarchyType.INDEPENDENT, StandardField.AUTHOR, "author2", true, ',', false));
@@ -173,7 +175,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void findMatchesMultipleEntries() throws Exception {
+    void findMatchesMultipleEntries() {
         GroupTreeNode parent = getNodeInSimpleTree();
         GroupTreeNode node = parent.addSubgroup(
                 new WordKeywordGroup("node", GroupHierarchyType.INDEPENDENT, StandardField.AUTHOR, "author1", true, ',', false));
@@ -181,7 +183,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void findMatchesWorksForRefiningGroups() throws Exception {
+    void findMatchesWorksForRefiningGroups() {
         GroupTreeNode grandParent = getNodeInSimpleTree();
         GroupTreeNode parent = grandParent.addSubgroup(
                 new WordKeywordGroup("node", GroupHierarchyType.INDEPENDENT, StandardField.AUTHOR, "author2", true, ',', false));
@@ -191,7 +193,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void findMatchesWorksForHierarchyOfIndependentGroups() throws Exception {
+    void findMatchesWorksForHierarchyOfIndependentGroups() {
         GroupTreeNode grandParent = getNodeInSimpleTree();
         GroupTreeNode parent = grandParent.addSubgroup(
                 new WordKeywordGroup("node", GroupHierarchyType.INDEPENDENT, StandardField.AUTHOR, "author2", true, ',', false));
@@ -201,7 +203,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void setGroupChangesUnderlyingGroup() throws Exception {
+    void setGroupChangesUnderlyingGroup() {
         GroupTreeNode node = getNodeInSimpleTree();
         AbstractGroup newGroup = new ExplicitGroup("NewGroup", GroupHierarchyType.INDEPENDENT, ',');
 
@@ -211,7 +213,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void setGroupAddsPreviousAssignmentsExplicitToExplicit() throws Exception {
+    void setGroupAddsPreviousAssignmentsExplicitToExplicit() {
         ExplicitGroup oldGroup = new ExplicitGroup("OldGroup", GroupHierarchyType.INDEPENDENT, ',');
         oldGroup.add(entry);
         GroupTreeNode node = GroupTreeNode.fromGroup(oldGroup);
@@ -223,7 +225,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void setGroupWithFalseDoesNotAddsPreviousAssignments() throws Exception {
+    void setGroupWithFalseDoesNotAddsPreviousAssignments() {
         ExplicitGroup oldGroup = new ExplicitGroup("OldGroup", GroupHierarchyType.INDEPENDENT, ',');
         oldGroup.add(entry);
         GroupTreeNode node = GroupTreeNode.fromGroup(oldGroup);
@@ -235,7 +237,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void setGroupAddsOnlyPreviousAssignments() throws Exception {
+    void setGroupAddsOnlyPreviousAssignments() {
         ExplicitGroup oldGroup = new ExplicitGroup("OldGroup", GroupHierarchyType.INDEPENDENT, ',');
         assertFalse(oldGroup.isMatch(entry));
         GroupTreeNode node = GroupTreeNode.fromGroup(oldGroup);
@@ -247,7 +249,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void setGroupExplicitToSearchDoesNotKeepPreviousAssignments() throws Exception {
+    void setGroupExplicitToSearchDoesNotKeepPreviousAssignments() {
         ExplicitGroup oldGroup = new ExplicitGroup("OldGroup", GroupHierarchyType.INDEPENDENT, ',');
         oldGroup.add(entry);
         GroupTreeNode node = GroupTreeNode.fromGroup(oldGroup);
@@ -259,7 +261,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void setGroupExplicitToExplicitIsRenameAndSoRemovesPreviousAssignment() throws Exception {
+    void setGroupExplicitToExplicitIsRenameAndSoRemovesPreviousAssignment() {
         ExplicitGroup oldGroup = new ExplicitGroup("OldGroup", GroupHierarchyType.INDEPENDENT, ',');
         oldGroup.add(entry);
         GroupTreeNode node = GroupTreeNode.fromGroup(oldGroup);
@@ -271,7 +273,7 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void getChildByPathFindsCorrectChildInSecondLevel() throws Exception {
+    void getChildByPathFindsCorrectChildInSecondLevel() {
         GroupTreeNode root = getRoot();
         GroupTreeNode child = getNodeInSimpleTree(root);
 
@@ -279,9 +281,67 @@ public class GroupTreeNodeTest {
     }
 
     @Test
-    void getPathSimpleTree() throws Exception {
+    void getChildByPathDoesNotFindChildWhenInvalidPath() {
+        GroupTreeNode root = getRoot();
+
+        // use side effect of method, which builds the group tree
+        getNodeInSimpleTree(root);
+
+        assertEquals(Optional.empty(), root.getChildByPath("ExplicitParent > ExplicitChildNode"));
+    }
+
+    @Test
+    void getPathSimpleTree() {
         GroupTreeNode node = getNodeInSimpleTree();
 
         assertEquals("ExplicitParent > ExplicitNode", node.getPath());
+    }
+
+    @Test
+    void onlyRootAndChildNodeContainAtLeastOneEntry() {
+        GroupTreeNode rootNode = getRoot();
+        rootNode.addSubgroup(new ExplicitGroup("ExplicitA", GroupHierarchyType.INCLUDING, ','));
+        GroupTreeNode parent = rootNode
+                .addSubgroup(new ExplicitGroup("ExplicitParent", GroupHierarchyType.INDEPENDENT, ','));
+        GroupTreeNode child = parent.addSubgroup(new ExplicitGroup("ExplicitNode", GroupHierarchyType.REFINING, ','));
+
+        BibEntry newEntry = new BibEntry().withField(StandardField.AUTHOR, "Stephen King");
+        child.addEntriesToGroup(Collections.singletonList(newEntry));
+        entries.add(newEntry);
+
+        assertEquals(rootNode.getContainingGroups(entries, false), Arrays.asList(rootNode, child));
+    }
+
+    @Test
+    void onlySubgroupsContainAllEntries() {
+        GroupTreeNode rootNode = getRoot();
+        rootNode.addSubgroup(new ExplicitGroup("ExplicitA", GroupHierarchyType.INCLUDING, ','));
+        GroupTreeNode parent = rootNode
+                .addSubgroup(new ExplicitGroup("ExplicitParent", GroupHierarchyType.INDEPENDENT, ','));
+        GroupTreeNode firstChild = parent.addSubgroup(new ExplicitGroup("ExplicitNode", GroupHierarchyType.REFINING, ','));
+        GroupTreeNode secondChild = parent.addSubgroup(new ExplicitGroup("ExplicitSecondNode", GroupHierarchyType.REFINING, ','));
+        GroupTreeNode grandChild = secondChild.addSubgroup(new ExplicitGroup("ExplicitNodeThirdLevel", GroupHierarchyType.REFINING, ','));
+
+        parent.addEntriesToGroup(Collections.singletonList(entry));
+        firstChild.addEntriesToGroup(entries);
+        secondChild.addEntriesToGroup(entries);
+        grandChild.addEntriesToGroup(entries);
+        assertEquals(parent.getContainingGroups(entries, true), Arrays.asList(firstChild, secondChild, grandChild));
+    }
+
+    @Test
+    void addEntriesToGroupWorksNotForGroupsNotSupportingExplicitAddingOfEntries() {
+        GroupTreeNode searchGroup = new GroupTreeNode(new SearchGroup("Search A", GroupHierarchyType.INCLUDING, "searchExpression", true, false));
+        List<FieldChange> fieldChanges = searchGroup.addEntriesToGroup(entries);
+
+        assertEquals(Collections.emptyList(), fieldChanges);
+    }
+
+    @Test
+    void removeEntriesFromGroupWorksNotForGroupsNotSupportingExplicitRemovalOfEntries() {
+        GroupTreeNode searchGroup = new GroupTreeNode(new SearchGroup("Search A", GroupHierarchyType.INCLUDING, "searchExpression", true, false));
+        List<FieldChange> fieldChanges = searchGroup.removeEntriesFromGroup(entries);
+
+        assertEquals(Collections.emptyList(), fieldChanges);
     }
 }
